B+Refactor: Move intro and settingspage into their feature folders
