+++ lib/features/settings/pages/SettingsPage.dart
@@ -1,8 +1,9 @@
-import 'package:decentproof/pages/settingspage/uiblocks/DevNetSwitch.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:decentproof/constants.dart';
 import 'package:url_launcher/url_launcher_string.dart';
+
+import '../uiblocks/DevNetSwitch.dart';
 
 class SettingsPage extends StatelessWidget {
   const SettingsPage({Key? key}) : super(key: key);
